v2.1.2 - Update feed initialization to use latest publish dates
Updated `FeedProcessor.cs` to retrieve and store the latest publish dates for both YouTube and RSS feeds. For YouTube feeds, `CheckFeedForUpdatesAsync` is now used to get the latest post's publish date. For RSS feeds, `CheckRssFeedForUpdatesAsync` is used to get the maximum publish date from the posts. These dates are stored in `youtubeFeedData` and `rssFeedData` respectively.